# mist_api_v2/openapi/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: api@mist.io
    name: Mist API Team
  title: Mist API
  version: 2.0.0
  x-cli-name: Mist CLI
servers:
- description: Mist Hosted Service
  url: https://mist.io/
paths:
...
  /api/v2/machines:
    get:
      description: List machines owned by the active org. READ permission required
        on machine & cloud.
      operationId: list_machines
      parameters:
      - examples:
          Filter by cloud id:
            value: 0194030499e74b02bdf68fa7130fb0b2
          Filter by cloud title:
            value: OSP10-new
          Filter by cloud provider:
            value: provider:digitalocean
        explode: true
        in: query
        name: cloud
        required: false
        schema:
          type: string
        style: form
      - description: Only return results matching search filter
        examples:
          Filter by machine type:
            value: state:running
        explode: true
        in: query
        name: search
        required: false
        schema:
          type: string
        style: form
      - description: Order results by
        examples:
          Sort by name, descending:
            value: -name
        explode: true
        in: query
        name: sort
        required: false
        schema:
          type: string
        style: form
      - description: Start results from index or id
        examples:
          Start from the 50th result:
            value: "50"
          Start from item with id xyz:
            value: xyz
        explode: true
        in: query
        name: start
        required: false
        schema:
          type: string
        style: form
      - description: Limit number of results, 1000 max
        explode: true
        in: query
        name: limit
        required: false
        schema:
          type: integer
        style: form
      - description: Only return these fields
        examples:
          Only return id:
            value: id
          Return id, name & owned_by:
            value: id,name,owned_by
        explode: true
        in: query
        name: only
        required: false
        schema:
          type: string
        style: form
      - description: Dereference foreign keys
        examples:
          '(default) Automatically replace all foreign key ids with referenced resource name ':
            value: auto
          Do not dereference anything, just return ids:
            value: none
          Dereference cloud id to cloud provider:
            value: cloud:provider
          Dereference cloud by provider, creator by email, owner by name:
            value: cloud:provider, creator:email, owned_by:name
        explode: true
        in: query
        name: deref
        required: false
        schema:
          default: auto
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMachinesResponse'
          description: List machines response
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
      security:
      - ApiKeyAuth:
        - read:machine
        - read:cloud
      - CookieAuth:
        - read:machine
        - read:cloud
      summary: List machines
      tags:
      - machines
      x-cli-table-default:
      - name
      - cloud
      - state
      - tags
      x-cli-table-extended:
      - id
      - name
      - state
      - cloud
      - external_id
      - public_ips
      - tags
      - owned_by
      - created_by
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
    post:
      description: Creates one or more machines on the specified cloud. If async is
        true, a jobId will be returned. READ permission required on cloud. CREATE_RESOURCES
        permission required on cloud. READ permission required on location. CREATE_RESOURCES
        permission required on location. CREATE permission required on machine. RUN
        permission required on script. READ permission required on key.
      operationId: create_machine
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMachineRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMachineResponse'
          description: Machine provisioning started
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
      security:
      - ApiKeyAuth:
        - create:machine
        - read:cloud
        - create_resources:cloud
        - read:location
        - create_resources:location
        - read:image
        - create_resources:image
        - read:script
        - read:key
      - CookieAuth:
        - create:machine
        - read:cloud
        - create_resources:cloud
        - read:location
        - create_resources:location
        - read:image
        - create_resources:image
        - read:script
        - read:key
      summary: Create machine
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
  /api/v2/machines/{machine}:
    get:
      description: Get details about target machine
      operationId: get_machine
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      - description: Only return these fields
        examples:
          Only return id:
            value: id
          Return id, name & owned_by:
            value: id,name,owned_by
        explode: true
        in: query
        name: only
        required: false
        schema:
          type: string
        style: form
      - description: Dereference foreign keys
        examples:
          '(default) Automatically replace all foreign key ids with referenced resource name ':
            value: auto
          Do not dereference anything, just return ids:
            value: none
          Dereference cloud id to cloud provider:
            value: cloud:provider
          Dereference cloud by provider, creator by email, owner by name:
            value: cloud:provider, creator:email, owned_by:name
        explode: true
        in: query
        name: deref
        required: false
        schema:
          default: auto
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMachineResponse'
          description: Machine returned successfully
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine not found
      security:
      - ApiKeyAuth:
        - read:machine
        - read:cloud
      - CookieAuth:
        - read:machine
        - read:cloud
      summary: Get machine
      tags:
      - machines
      x-cli-table-default:
      - name
      - cloud
      - state
      - tags
      x-cli-table-extended:
      - id
      - name
      - state
      - cloud
      - external_id
      - public_ips
      - tags
      - owned_by
      - created_by
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
    put:
      description: Edit target machine
      operationId: edit_machine
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      - description: New machine name
        explode: true
        in: query
        name: name
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          description: Machine successfully updated
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine not found
      security:
      - ApiKeyAuth:
        - edit:machine
      - CookieAuth:
        - edit:machine
      summary: Edit machine
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
  /api/v2/machines/{machine}/actions/associate-key:
    put:
      description: Associate a key with a machine.
      operationId: associate_key
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyMachineAssociation'
      responses:
        "200":
          description: Association successful
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine or Key not found
        "422":
          description: Action not supported on target machine
        "503":
          description: Could not connect to target machine
      security:
      - ApiKeyAuth:
        - associate_key:machine
        - read:cloud
        - read_private:key
      - CookieAuth:
        - associate_key:machine
        - read:cloud
        - read_private:key
      summary: Associate a key with a machine
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
  /api/v2/machines/{machine}/actions/clone:
    post:
      description: Clone target machine
      operationId: clone_machine
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Machine clone issued successfully
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine not found
        "422":
          description: Action not supported on target machine
      security:
      - ApiKeyAuth:
        - clone:machine
      - CookieAuth:
        - clone:machine
      summary: Clone machine
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
  /api/v2/machines/{machine}/actions/console:
    post:
      description: Open VNC console on target machine
      operationId: console
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Machine console opened successfully
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine not found
        "422":
          description: Action not supported on target machine
      security:
      - ApiKeyAuth:
        - console:machine
      - CookieAuth:
        - console:machine
      summary: Open console
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
  /api/v2/machines/{machine}/actions/destroy:
    post:
      description: Destroy target machine
      operationId: destroy_machine
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Machine destroy issued successfully
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine not found
        "422":
          description: Action not supported on target machine
      security:
      - ApiKeyAuth:
        - destroy:machine
      - CookieAuth:
        - destroy:machine
      summary: Destroy machine
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
  /api/v2/machines/{machine}/actions/disassociate-key:
    delete:
      description: Disassociate a key from a machine.
      operationId: disassociate_key
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyMachineDisassociation'
      responses:
        "200":
          description: Disassociation successful
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine or Key not found
        "422":
          description: Action not supported on target machine
      security:
      - ApiKeyAuth:
        - disassociate_key:machine
        - read:cloud
      - CookieAuth:
        - disassociate_key:machine
        - read:cloud
      summary: Disassociate a key from a machine
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
  /api/v2/machines/{machine}/actions/expose:
    post:
      description: Expose target machine
      operationId: expose_machine
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Machine expose issued successfully
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine not found
        "422":
          description: Action not supported on target machine
      security:
      - ApiKeyAuth:
        - expose:machine
      - CookieAuth:
        - expose:machine
      summary: Expose machine
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
  /api/v2/machines/{machine}/actions/reboot:
    post:
      description: Reboot target machine
      operationId: reboot_machine
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Machine reboot issued successfully
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine not found
        "422":
          description: Action not supported on target machine
      security:
      - ApiKeyAuth:
        - reboot:machine
      - CookieAuth:
        - reboot:machine
      summary: Reboot machine
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
  /api/v2/machines/{machine}/actions/rename:
    post:
      description: Rename target machine
      operationId: rename_machine
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Machine renamed successfully
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine not found
        "422":
          description: Action not supported on target machine
      security:
      - ApiKeyAuth:
        - rename:machine
      - CookieAuth:
        - rename:machine
      summary: Rename machine
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
  /api/v2/machines/{machine}/actions/resize:
    post:
      description: Resize target machine
      operationId: resize_machine
      parameters:
      - explode: false
        in: path
        name: machine
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Machine resize issued successfully
        "401":
          description: Authentication failed
        "403":
          description: You are not authorized to perform this action
        "404":
          description: Machine not found
        "422":
          description: Action not supported on target machine
      security:
      - ApiKeyAuth:
        - resize:machine
      - CookieAuth:
        - resize:machine
      summary: Resize machine
      tags:
      - machines
      x-openapi-router-controller: mist_api_v2.controllers.machines_controller
...
components:
...
  schemas:
...
    CreateMachineResponse:
      oneOf:
      - $ref: '#/components/schemas/CreateMachineResponse_oneOf'
      - $ref: '#/components/schemas/CreateMachineResponse_oneOf_1'
      type: object
...
    KeyMachineDisassociation:
      example:
        key: key
      properties:
        key:
          description: Name or ID of the SSH key to disassociate
          type: string
      type: object
...
    ListMachinesResponse:
      example:
        data:
        - cloud: cloud
          default: true
          name: name
          external_id: external_id
          owned_by: owned_by
          id: id
          created_by: created_by
          tags: '{}'
        - cloud: cloud
          default: true
          name: name
          external_id: external_id
          owned_by: owned_by
          id: id
          created_by: created_by
          tags: '{}'
        meta:
          total: 0
          start: 1
          sort: sort
          returned: 6
      properties:
        data:
          items:
            $ref: '#/components/schemas/Machine'
          type: array
        meta:
          $ref: '#/components/schemas/ResponseMetadata'
      type: object
...
    CreateMachineResponse_oneOf:
      properties:
        jobId:
          type: string
        plan:
          type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
      x-apikeyInfoFunc: mist_api_v2.controllers.security_controller_.info_from_ApiKeyAuth
    CookieAuth:
      in: cookie
      name: session.id
      type: apiKey
      x-apikeyInfoFunc: mist_api_v2.controllers.security_controller_.info_from_CookieAuth
...